+++ packages/app-utils/src/plugins/remote-component.config.ts
@@ -20,6 +20,7 @@ import * as DeephavenJsapiUtils from '@deephaven/jsapi-utils';
 import DeephavenLog from '@deephaven/log';
 import * as DeephavenReactHooks from '@deephaven/react-hooks';
 import * as DeephavenPlugin from '@deephaven/plugin';
+import * as DeephavenCorePlugins from '@deephaven/dashboard-core-plugins';
 
 // eslint-disable-next-line import/prefer-default-export
 export const resolve = {
@@ -40,4 +41,5 @@ export const resolve = {
   '@deephaven/log': DeephavenLog,
   '@deephaven/plugin': DeephavenPlugin,
   '@deephaven/react-hooks': DeephavenReactHooks,
+  '@deephaven/dashboard-core-plugins': DeephavenCorePlugins,
 };
