+++ packages/dashboard-core-plugins/src/ChartPluginConfig.ts
@@ -1,12 +1,13 @@
 import { PluginType, type WidgetPlugin } from '@deephaven/plugin';
 import { vsGraph } from '@deephaven/icons';
-import { ChartPlugin, ChartPanelPlugin } from './ChartPlugin';
+import { ChartWidgetPlugin } from './ChartWidgetPlugin';
+import { ChartPanelPlugin } from './ChartPanelPlugin';
 
 const ChartPluginConfig: WidgetPlugin = {
   name: 'ChartPanel',
   title: 'Chart',
   type: PluginType.WIDGET_PLUGIN,
-  component: ChartPlugin,
+  component: ChartWidgetPlugin,
   panelComponent: ChartPanelPlugin,
   supportedTypes: 'Figure',
   icon: vsGraph,
