+++ packages/dashboard-core-plugins/src/GridPluginConfig.ts
@@ -1,12 +1,13 @@
 import { PluginType, type WidgetPlugin } from '@deephaven/plugin';
 import { dhTable } from '@deephaven/icons';
-import { GridPlugin, GridPanelPlugin } from './GridPlugin';
+import { GridWidgetPlugin } from './GridWidgetPlugin';
+import { GridPanelPlugin } from './GridPanelPlugin';
 
 const GridPluginConfig: WidgetPlugin = {
   name: 'IrisGridPanel',
   title: 'Table',
   type: PluginType.WIDGET_PLUGIN,
-  component: GridPlugin,
+  component: GridWidgetPlugin,
   panelComponent: GridPanelPlugin,
   supportedTypes: ['Table', 'TreeTable', 'HierarchicalTable'],
   icon: dhTable,
