+++ packages/dashboard-core-plugins/src/PandasWidgetPlugin.tsx
@@ -1,5 +1,5 @@
-import { forwardRef, useCallback, useEffect, useState } from 'react';
-import { WidgetComponentProps, WidgetPanelProps } from '@deephaven/plugin';
+import { useCallback, useEffect, useState } from 'react';
+import { WidgetComponentProps } from '@deephaven/plugin';
 import { type Table } from '@deephaven/jsapi-types';
 import IrisGrid, {
   IrisGridModelFactory,
@@ -7,10 +7,11 @@ import IrisGrid, {
 } from '@deephaven/iris-grid';
 import { useApi } from '@deephaven/jsapi-bootstrap';
 import { LoadingOverlay } from '@deephaven/components';
-import { PandasPanel, PandasReloadButton } from './panels';
-import useHydrateGrid from './useHydrateGrid';
+import { PandasReloadButton } from './panels';
 
-export function PandasPlugin(props: WidgetComponentProps): JSX.Element | null {
+export function PandasWidgetPlugin(
+  props: WidgetComponentProps
+): JSX.Element | null {
   const dh = useApi();
   const [model, setModel] = useState<IrisGridModel>();
   const [isLoading, setIsLoading] = useState(true);
@@ -61,19 +62,4 @@ export function PandasPlugin(props: WidgetComponentProps): JSX.Element | null {
   );
 }
 
-export const PandasPanelPlugin = forwardRef(
-  (props: WidgetPanelProps, ref: React.Ref<PandasPanel>) => {
-    const { localDashboardId, fetch } = props;
-    const hydratedProps = useHydrateGrid(
-      fetch as unknown as () => Promise<Table>,
-      localDashboardId
-    );
-
-    return (
-      // eslint-disable-next-line react/jsx-props-no-spreading
-      <PandasPanel ref={ref} {...props} {...hydratedProps} />
-    );
-  }
-);
-
-PandasPanelPlugin.displayName = 'PandasPanelPlugin';
+export default PandasWidgetPlugin;
