+++ packages/dashboard-core-plugins/src/index.ts
@@ -1,17 +1,20 @@
-export * from './ChartPlugin';
+export { default as ChartPanelPlugin } from './ChartPanelPlugin';
+export { default as ChartWidgetPlugin } from './ChartWidgetPlugin';
 export { default as ChartPluginConfig } from './ChartPluginConfig';
 export { default as ChartBuilderPlugin } from './ChartBuilderPlugin';
 export { default as ChartBuilderPluginConfig } from './ChartBuilderPluginConfig';
 export { default as ConsolePlugin } from './ConsolePlugin';
 export { default as FilterPlugin } from './FilterPlugin';
 export { default as FilterPluginConfig } from './FilterPluginConfig';
-export * from './GridPlugin';
+export { default as GridPanelPlugin } from './GridPanelPlugin';
+export { default as GridWidgetPlugin } from './GridWidgetPlugin';
 export { default as GridPluginConfig } from './GridPluginConfig';
 export { default as LinkerPlugin } from './LinkerPlugin';
 export { default as LinkerPluginConfig } from './LinkerPluginConfig';
 export { default as MarkdownPlugin } from './MarkdownPlugin';
 export { default as MarkdownPluginConfig } from './MarkdownPluginConfig';
-export * from './PandasPlugin';
+export { default as PandasPanelPlugin } from './PandasPanelPlugin';
+export { default as PandasWidgetPlugin } from './PandasWidgetPlugin';
 export { default as PandasPluginConfig } from './PandasPluginConfig';
 export { default as WidgetLoaderPlugin } from './WidgetLoaderPlugin';
 export { default as WidgetLoaderPluginConfig } from './WidgetLoaderPluginConfig';
